TRIX-122 allowing multiple nameservers and somewhat make the magic happen. not sure regarding the order/priority

+++ daemon/plugins/boot/network/default.py
@@ -74,11 +74,14 @@ class Plugin():
 
     dns = """
         cd /sysroot
-        echo -n '' > etc/resolv.conf
+        if [ ! -e tmp/resolv.clear ]; then
+            echo -n '' > etc/resolv.conf
+            search=$(echo $SEARCH | awk '{gsub(/,/, " "); print}')
+            echo "search $search" >> etc/resolv.conf
+            > tmp/resolv.clear
+        fi
         for server in $(echo $NAMESERVER|tr ',' ' '); do
             echo "nameserver $server" >> etc/resolv.conf
         done
-        search=$(echo $SEARCH | awk '{gsub(/,/, " "); print}')
-        echo "search $search" >> etc/resolv.conf
     """
